fix: unit tests

+++ home/home_data/src/main/java/com/acidtango/home_data/network/ReceiptsMapper.kt
@@ -13,7 +13,7 @@ fun ReceiptDetailNetwork.toDomain() = ReceiptDetail(
     name = this.name,
     favourite = this.favourite,
     favouriteAmount = this.favouriteAmount,
-    preparationTime = this.preparationTime
+    preparationTime = this.preparationTime.toLong()
 )
 
 fun MetaNetwork.toDomain() = Meta(totalItems = this.totalItems)
